Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -20,7 +20,7 @@ int main()
         sum += d[i].temp;
     }
 
-    float avg = sum / n;
+    float avg = sum / (float)n;
     printf("Average Temperature: %.2f\n", avg);
 
     return 0;
